+++ app/Http/Controllers/Staff/LeadController.php
@@ -15,11 +15,12 @@ use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Log;
 
 class LeadController extends Controller {
+
     /**
      * Display a listing of the resource.
      */
     public function index() {
-        $this->authorize('staff.leads.index');
+        $this->authorize('staff.leads');
 
         $leads = Client::with(['user', 'partner', 'partner.user'])->where('active', 0)->where('status', '!=', Client::QUALIFIED)->get();
 
@@ -49,6 +50,8 @@ class LeadController extends Controller {
      * Store a newly created resource in storage.
      */
     public function store(StoreLead $request) {
+        $this->authorize('staff.leads');
+
         $validatedData = $request->validated();
         $validatedData['password'] = Hash::make($validatedData['password']);
 
@@ -97,6 +100,8 @@ class LeadController extends Controller {
      * Update the specified resource in storage.
      */
     public function update(UpdateLead $request, string $id) {
+        $this->authorize('staff.leads');
+
         $lead = Client::findOrFail($id);
         $validatedData = $request->validated();
 
@@ -127,6 +132,8 @@ class LeadController extends Controller {
      * Remove the specified resource from storage.
      */
     public function destroy(string $id) {
+        $this->authorize('staff.leads');
+
         $lead = Client::findOrFail($id);
         $lead->user()->delete();
         $lead->delete();
@@ -136,6 +143,8 @@ class LeadController extends Controller {
 
     // Update lead status
     public function updateLeadStatus(Request $request, string $id) {
+        $this->authorize('staff.leads');
+
         try {
             $lead = Client::findOrFail($id);
             $lead->status = $request->status;
@@ -149,6 +158,8 @@ class LeadController extends Controller {
 
     // Fetch lead by ID
     public function fetchLead(string $id) {
+        $this->authorize('staff.leads');
+
         $lead = Client::with(['user', 'partner', 'partner.user'])
             ->findOrFail($id);
         return response()->json([
